First Commit
This is my original capstone project for my final semester at Mohawk.
I've removed some files and code that point to the Azure service I was
using that is no longer in service.  This application won't run as is
since it needs access to the database, which is not available, but is
available here for code review and safe-keeping.